+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/ProcessorConfigurationService.java
@@ -10,28 +10,6 @@
 
 package com.liaison.mailbox.service.core;
 
-import java.io.IOException;
-import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityTransaction;
-import javax.persistence.NoResultException;
-import javax.ws.rs.core.Response;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
 import com.liaison.commons.jpa.DAOUtil;
 import com.liaison.commons.util.UUIDGen;
 import com.liaison.mailbox.MailBoxConstants;
@@ -67,6 +45,7 @@ import com.liaison.mailbox.service.core.processor.MailBoxProcessorI;
 import com.liaison.mailbox.service.dto.GenericSearchFilterDTO;
 import com.liaison.mailbox.service.dto.HTTPListenerHelperDTO;
 import com.liaison.mailbox.service.dto.ResponseDTO;
+import com.liaison.mailbox.service.dto.configuration.DatacenterDTO;
 import com.liaison.mailbox.service.dto.configuration.DynamicPropertiesDTO;
 import com.liaison.mailbox.service.dto.configuration.MailBoxDTO;
 import com.liaison.mailbox.service.dto.configuration.ProcessorDTO;
@@ -79,6 +58,7 @@ import com.liaison.mailbox.service.dto.configuration.request.ReviseProcessorRequ
 import com.liaison.mailbox.service.dto.configuration.response.AddProcessorToMailboxResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.response.ClusterTypeResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.response.DeActivateProcessorResponseDTO;
+import com.liaison.mailbox.service.dto.configuration.response.GetDatacenterResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.response.GetProcessorIdResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.response.GetProcessorResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.response.ProcessorResponseDTO;
@@ -90,6 +70,29 @@ import com.liaison.mailbox.service.util.MailBoxUtil;
 import com.liaison.mailbox.service.util.ProcessorPropertyJsonMapper;
 import com.liaison.mailbox.service.validation.GenericValidator;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.codehaus.jackson.map.ObjectMapper;
+import org.codehaus.jackson.type.TypeReference;
+
+import javax.persistence.EntityManager;
+import javax.persistence.EntityTransaction;
+import javax.persistence.NoResultException;
+import javax.ws.rs.core.Response;
+
+import java.io.IOException;
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
 
 /**
  * Class which has Processor configuration related operations.
@@ -99,6 +102,7 @@ import com.liaison.mailbox.service.validation.GenericValidator;
 public class ProcessorConfigurationService {
 
 	private static final Logger LOGGER = LogManager.getLogger(ProcessorConfigurationService.class);
+	private static final String DATACENTER_PROCESSOR = "Datacenetres and Processors";
 
 	/**
 	 * Creates processor for the mailbox.
@@ -1238,6 +1242,101 @@ public class ProcessorConfigurationService {
     }
     
     /**
+     * This method support to processoraffinity This is to split the traffic
+     * for the processor that need to run different datacenters to support
+     * active active
+     * 
+     * @param request
+     */
+    public void supportProcessorAffinity(String request) {
+
+        LOGGER.debug("Enter into supportProcessorAffinity () ");
+        try {
+            
+            //retrieve the datacenetre and value is the % of processor that should run on that DC 
+            ObjectMapper mapper = new ObjectMapper();
+            Map<String, String> datacenterMap = mapper.readValue(request, new TypeReference<HashMap<String, Object>>() {});
+            //fetch processor count
+            ProcessorConfigurationDAO dao = new ProcessorConfigurationDAOBase();
+            long processorCount = dao.getProcessorCount();
+            if (0 == processorCount) {
+                LOGGER.info("No processor exist");
+                return;
+            }            
+            List<String> processedDC = new ArrayList<String>();            
+            for (String dc : datacenterMap.keySet()) {                
+                
+                processedDC.add(dc);
+                int readyToProcessCount = (int) Math.ceil((Integer.parseInt(datacenterMap.get(dc)) * processorCount) / 100);                
+                dao.updateDatacenter(dc, processedDC, readyToProcessCount);
+            }
+
+        } catch (Exception exception) {
+            throw new RuntimeException(exception);
+        }
+        
+        LOGGER.debug("Exit from supportProcessorAffinity () ");
+    }
+    
+    /**
+     * Method fetch the Datacenter's and corresponding processors details.
+     * 
+     * @return GetDatacenterResponseDTO
+     */
+    public GetDatacenterResponseDTO getDatacenters() {
+
+        LOGGER.debug("Enter into getDatacenters () ");
+
+        GetDatacenterResponseDTO serviceResponse = new GetDatacenterResponseDTO();
+
+        try {
+
+            ProcessorConfigurationDAO dao = new ProcessorConfigurationDAOBase();
+            List<ProcessorDTO> prsDTOList = null;
+            List<DatacenterDTO> dcDTOList = new ArrayList<DatacenterDTO>();
+            List<Processor> processors = null;
+            ProcessorDTO processorDTO = null;
+            // fetch the all datacenters
+            List<String> datacenters = dao.getAllDatacenters();
+            DatacenterDTO dcDTO = null;
+            for (String name : datacenters) {
+
+                if (null == name) {
+                    continue;
+                }
+                dcDTO = new DatacenterDTO();
+                dcDTO.setName(name);
+                // fetch the all Processor's
+                processors = dao.findProcessorsByDatacenter(name);
+                prsDTOList = new ArrayList<ProcessorDTO>();
+
+                for (Processor processor : processors) {
+                    processorDTO = new ProcessorDTO();
+                    processorDTO.copyFromEntity(processor, false);
+                    prsDTOList.add(processorDTO);
+                }
+
+                dcDTO.setProcessors(prsDTOList);
+                dcDTO.setTotalItems(prsDTOList.size());
+                dcDTOList.add(dcDTO);
+            }
+
+            serviceResponse.setDatacenetrs(dcDTOList);
+            serviceResponse.setResponse(new ResponseDTO(Messages.READ_SUCCESSFUL, DATACENTER_PROCESSOR,
+                    Messages.SUCCESS));
+
+            LOGGER.debug("Exit from getDatacenters () ");
+
+        } catch (Exception e) {
+            LOGGER.error(Messages.READ_OPERATION_FAILED.name(), e);
+            serviceResponse.setResponse(new ResponseDTO(Messages.READ_OPERATION_FAILED, DATACENTER_PROCESSOR,
+                    Messages.FAILURE, e.getMessage()));
+        }
+
+        return serviceResponse;
+    }
+
+    /**
      * Helper method validate the Processor type by deployment type
      * 
      * @param processorType
@@ -1315,5 +1414,12 @@ public class ProcessorConfigurationService {
             default:
                 return false;
         }
+    }
+
+    /**
+     * This method is used to update the process_dc to the current_Dc where the process_dc is null
+     */
+    public void updateProcessDc() {
+        new ProcessorConfigurationDAOBase().updateProcessDc();
     }
 }