+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 * GMB-948 Log the application logs to Kibana directly
 * GMB-1063 - Unable to receive the response payload in http sync request
 * GMB-912 - File writer replication retry if storage replication is not done.
+* GMB-1011 - Processor affinity support
 
 4.17.3
 ---------------